separated dom manip from fetch logic. updated package.json
