+++ src/index.js
@@ -1,26 +1,27 @@
 import 'bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './style.css';
+import domManip from './domManip';
 
-const search = document.getElementById('search');
-const loading = document.getElementById('loading');
-const input = document.getElementById('searchInput');
-const form = document.getElementById('form');
-const city = document.getElementById('city');
-const description = document.getElementById('description');
-const temp = document.getElementById('temp');
-const feel = document.getElementById('feel');
-const humidity = document.getElementById('humidity');
-const alert = document.getElementById('alert');
-const danger = document.getElementById('danger');
-const icon = document.getElementById('icon');
-const deg = document.getElementById('deg');
-const fahr = document.getElementById('fahr');
-const details1 = document.getElementById('details1');
-const details2 = document.getElementById('details2');
-const img = document.createElement('img');
-
-icon.appendChild(img);
+const dom = domManip();
+const {
+  search,
+  loading,
+  input,
+  form,
+  city,
+  description,
+  temp,
+  feel,
+  humidity,
+  alert,
+  danger,
+  deg,
+  fahr,
+  details1,
+  details2,
+  img,
+} = dom;
 
 const addUnit = (str, valTemp, valFeel) => {
   temp.innerHTML = valTemp;
@@ -95,7 +96,15 @@ const getData = async (val) => {
       }, 5000);
     }
   } catch (e) {
-    alert(`There is an error${e}`);
+    details1.style.display = 'none';
+    details2.style.display = 'none';
+    danger.style.display = 'block';
+    danger.innerHTML = `There is an error${e}`;
+    search.style.display = 'block';
+    loading.style.display = 'none';
+    setTimeout(() => {
+      danger.style.display = 'none';
+    }, 5000);
   }
 };
